+++ src/main/java/com/nextdaydelivery/ai_response/domain/entity/AiResponse.java
@@ -1,6 +1,6 @@
 package com.nextdaydelivery.ai_response.domain.entity;
 
-import com.nextdaydelivery.global.domain.CreatedAuditEntity;
+import com.nextdaydelivery.global.domain.entity.CreatedAuditEntity;
 import com.nextdaydelivery.user.domain.entity.User;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
