+++ src/main/java/com/nextdaydelivery/global/domain/error/GlobalErrorCode.java
@@ -1,4 +1,4 @@
-package com.nextdaydelivery.global.domain;
+package com.nextdaydelivery.global.domain.error;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
