+++ src/main/java/com/nextdaydelivery/global/dto/CommonResponse.java
@@ -3,7 +3,7 @@ package com.nextdaydelivery.global.dto;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.nextdaydelivery.global.domain.ErrorCode;
+import com.nextdaydelivery.global.domain.error.ErrorCode;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.http.HttpStatus;
