+++ src/main/java/com/nextdaydelivery/payment/domain/entity/Payment.java
@@ -1,6 +1,6 @@
 package com.nextdaydelivery.payment.domain.entity;
 
-import com.nextdaydelivery.global.domain.CreatedAuditEntity;
+import com.nextdaydelivery.global.domain.entity.CreatedAuditEntity;
 import com.nextdaydelivery.order.domain.entity.Order;
 import com.nextdaydelivery.payment.domain.enums.PaymentMethod;
 import com.nextdaydelivery.payment.domain.enums.PaymentStatus;
